refactor: 💡 rename test cookie route
