+++ app/routes/__init__.py
@@ -1,5 +1,5 @@
 from .my_pages import MyPages
 from .my_pages_status import MyPagesStatus
 from .applications import Applications
-from .testcookie import TestCookie
+from .test_cookie import TestCookie
 from .index import Index
